Add customColor prop on HorizontalNav component
